Move type check helpers into the `internal` library
