+++ packages/windows_foundation/lib/internal.dart
@@ -14,6 +14,7 @@ export 'src/internal/hstring_array.dart';
 export 'src/internal/iids.dart';
 export 'src/internal/int_array.dart';
 export 'src/internal/ipropertyvalue_helpers.dart';
+export 'src/internal/type_check_helpers.dart';
 export 'src/ipropertyvalue.dart';
 export 'src/ireference.dart';
 export 'src/propertyvalue.dart';
