+++ packages/windows_foundation/lib/src/collections/iiterator.dart
@@ -7,6 +7,7 @@ import 'dart:ffi';
 import 'package:ffi/ffi.dart';
 import 'package:win32/win32.dart';
 
+import '../../internal.dart';
 import '../helpers.dart';
 import '../iinspectable.dart';
 import '../types.dart';
